Move DataDescription component into separate file
